+++ tests/Factory/Psr17/Psr17FactoryProviderTest.php
@@ -18,6 +18,7 @@ class Psr17FactoryProviderTest extends TestCase
     /**
      * @runInSeparateProcess - Psr17FactoryProvider::setFactories breaks other tests
      */
+    #[\PHPUnit\Framework\Attributes\RunInSeparateProcess]
     public function testGetSetFactories()
     {
         Psr17FactoryProvider::setFactories([]);
@@ -29,6 +30,7 @@ class Psr17FactoryProviderTest extends TestCase
     /**
      * @runInSeparateProcess - Psr17FactoryProvider::setFactories breaks other tests
      */
+    #[\PHPUnit\Framework\Attributes\RunInSeparateProcess]
     public function testAddFactory()
     {
         Psr17FactoryProvider::setFactories(['Factory 1']);