+++ tests/Routing/FastRouteDispatcherTest.php
@@ -23,6 +23,7 @@ class FastRouteDispatcherTest extends TestCase
     /**
      * @dataProvider provideFoundDispatchCases
      */
+    #[\PHPUnit\Framework\Attributes\DataProvider('provideFoundDispatchCases')]
     public function testFoundDispatches($method, $uri, $callback, $handler, $argDict)
     {
         /** @var FastRouteDispatcher $dispatcher */
@@ -59,6 +60,7 @@ class FastRouteDispatcherTest extends TestCase
     /**
      * @dataProvider provideNotFoundDispatchCases
      */
+    #[\PHPUnit\Framework\Attributes\DataProvider('provideNotFoundDispatchCases')]
     public function testNotFoundDispatches($method, $uri, $callback)
     {
         /** @var FastRouteDispatcher $dispatcher */
@@ -74,8 +76,10 @@ class FastRouteDispatcherTest extends TestCase
      * @param $method
      * @param $uri
      * @param $callback
+     * @param $allowedMethods
      */
-    public function testMethodNotAllowedDispatches($method, $uri, $callback)
+    #[\PHPUnit\Framework\Attributes\DataProvider('provideMethodNotAllowedDispatchCases')]
+    public function testMethodNotAllowedDispatches($method, $uri, $callback, $allowedMethods)
     {
         /** @var FastRouteDispatcher $dispatcher */
         $dispatcher = simpleDispatcher($callback, $this->generateDispatcherOptions());
@@ -92,6 +96,7 @@ class FastRouteDispatcherTest extends TestCase
      * @param $callback
      * @param $allowedMethods
      */
+    #[\PHPUnit\Framework\Attributes\DataProvider('provideMethodNotAllowedDispatchCases')]
     public function testGetAllowedMethods($method, $uri, $callback, $allowedMethods)
     {
         /** @var FastRouteDispatcher $dispatcher */
@@ -160,7 +165,7 @@ class FastRouteDispatcherTest extends TestCase
         }, $this->generateDispatcherOptions());
     }
 
-    public function provideFoundDispatchCases()
+    public static function provideFoundDispatchCases()
     {
         $cases = [];
 
@@ -459,7 +464,7 @@ class FastRouteDispatcherTest extends TestCase
         return $cases;
     }
 
-    public function provideNotFoundDispatchCases()
+    public static function provideNotFoundDispatchCases()
     {
         $cases = [];
 
@@ -536,7 +541,7 @@ class FastRouteDispatcherTest extends TestCase
         return $cases;
     }
 
-    public function provideMethodNotAllowedDispatchCases()
+    public static function provideMethodNotAllowedDispatchCases()
     {
         $cases = [];